updating genrl and vae code
