--- emote/nn/misc.py
@@ -1,44 +1,0 @@
-import torch.nn.functional as F
-
-from torch import Tensor, nn
-
-
-class FrozenBatchNorm1d(nn.BatchNorm1d):
-    def __init__(
-        self,
-        n_features: int,
-        eps: float = 1e-5,
-        momentum: float = 0.1,
-        affine: bool = True,
-        track_running_stats: bool = True,
-        device=None,
-        dtype=None,
-    ) -> None:
-        factory_kwargs = {"device": device, "dtype": dtype}
-        super().__init__(
-            n_features, eps, momentum, affine, track_running_stats, **factory_kwargs
-        )
-
-    def forward(self, input: Tensor) -> Tensor:
-        self._check_input_dim(input)
-        if self.momentum is None:
-            exponential_average_factor = 0.0
-        else:
-            exponential_average_factor = self.momentum
-        bn_training = False
-        return F.batch_norm(
-            input,
-            self.running_mean,
-            self.running_var,
-            self.weight,
-            self.bias,
-            bn_training,
-            exponential_average_factor,
-            self.eps,
-        )
-
-    def _check_input_dim(self, input):
-        if input.dim() != 2 and input.dim() != 3:
-            raise ValueError(
-                "expected 2D or 3D input (got {}D input)".format(input.dim())
-            )
